requirements.txt
torch
torchvision
pudb
pytorch-lightning
einops
regex
ftfy
cython
pillow
sklearn
tensorboard
PyWavelets
opencv-python